+++ app/src/main/java/com/example/scheduleapp/AddActivity.kt
@@ -5,26 +5,23 @@ import android.os.Bundle
 import android.view.View
 import android.widget.CalendarView
 import android.widget.EditText
+import java.text.SimpleDateFormat
 
 class AddActivity : AppCompatActivity(),
-    DatePickerDialogFragment.OnSelectedDateListener,
     TimePickerDialogFragment.OnSelectedTimeListener {
 
     //DBヘルパー
     private val helper = DatabaseHelper(this@AddActivity)
+    private var selectedDate = "";
 
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
         setContentView(R.layout.activity_add)
 
-        //日付入力欄の取得
-        val etDate = findViewById<EditText>(R.id.etDate)
-        //前画面から引き継いだ初期値を設定
-        etDate.setText(intent.getStringExtra("selectedDate"))
-        //選択時にダイアログを表示するためのリスナ設定
-        etDate.setOnClickListener {
-            showDatePickerDialog()
-        }
+        //選択されている日付の初期値を設定
+        val calendar = findViewById<CalendarView>(R.id.calendarView)
+        val sdf = SimpleDateFormat("yyyy/MM/dd")
+        selectedDate = sdf.format(calendar.date)
 
         //時間入力欄の取得
         val etTime = findViewById<EditText>(R.id.etTime)
@@ -33,17 +30,6 @@ class AddActivity : AppCompatActivity(),
             showTimePickerDialog()
         }
 
-    }
-
-    private fun showDatePickerDialog() {
-        val datePickerDialogFragment = DatePickerDialogFragment()
-        datePickerDialogFragment.show(supportFragmentManager, null)
-    }
-
-    override fun selectedDate(year: Int, month: Int, date: Int) {
-        val text = "$year/${month+1}/$date"
-        val etDate = findViewById<EditText>(R.id.etDate)
-        etDate.setText(text)
     }
 
     private fun showTimePickerDialog() {
@@ -62,8 +48,6 @@ class AddActivity : AppCompatActivity(),
      */
     fun onMakeButtonClick(view: View){
         //入力内容をDBに登録する処理
-        val etDate = findViewById<EditText>(R.id.etDate)
-        val etDateText = etDate.text.toString()
         val etTime = findViewById<EditText>(R.id.etTime)
         val etTimeText = etTime.text.toString()
         val etDesc = findViewById<EditText>(R.id.etDesc)
@@ -76,7 +60,7 @@ class AddActivity : AppCompatActivity(),
         //プリペアドステートメントの取得
         val stmt = db.compileStatement(sqlInsert)
         //変数のバインド
-        stmt.bindString(1,etDateText)
+        stmt.bindString(1,selectedDate)
         stmt.bindString(2,etTimeText)
         stmt.bindString(3,etDescText)
         //実行
@@ -90,5 +74,16 @@ class AddActivity : AppCompatActivity(),
         //ヘルパーオブジェクトの開放
         helper.close()
         super.onDestroy()
+    }
+
+    /**
+     * カレンダーの日付を変えたときのリスナ
+     */
+    private inner class DateChangeListener : CalendarView.OnDateChangeListener{
+        override fun onSelectedDayChange(calendarView: CalendarView, year: Int, month: Int, dayOfMonth: Int) {
+            // monthは0起算のため+1します。
+            val displayMonth = month + 1
+            selectedDate = "$year/$displayMonth/$dayOfMonth"
+        }
     }
 }
